+++ packages/cloudbase_core_nullsafety/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.0.3
+* feat: add CUSTOM and PHONE support.
+* 新增：加入自定义登录和手机号短信登录支持.
+
 ## 0.0.2
 * fix: requestId not null.
 * doc: add README.md